Final Cleanup
Notes: Based on the original code, this is calculating future one year interest based on the rules and rates given.
It does not compound interest, add interest into the balance or actually compound on a daily basis. If these are required features, they can be added simply but it seemed like that was not the intent of a two hour refactoring.